Fixes warnings about casts in C code.

+++ src/native/g722/net_java_sip_communicator_impl_neomedia_codec_audio_g722_JNIDecoder.c
@@ -1,6 +1,8 @@
 #include "net_java_sip_communicator_impl_neomedia_codec_audio_g722_JNIDecoder.h"
 
 #include <inttypes.h>
+#include <stdint.h>
+
 #include "telephony.h"
 #include "g722.h"
 
@@ -8,7 +10,7 @@ JNIEXPORT void JNICALL
 Java_net_java_sip_communicator_impl_neomedia_codec_audio_g722_JNIDecoder_g722_1decoder_1close
     (JNIEnv *jniEnv, jclass clazz, jlong decoder)
 {
-    g722_decode_state_t *d = (g722_decode_state_t *) decoder;
+    g722_decode_state_t *d = (g722_decode_state_t *) (intptr_t) decoder;
 
     g722_decode_release(d);
     g722_decode_free(d);
@@ -18,7 +20,7 @@ JNIEXPORT jlong JNICALL
 Java_net_java_sip_communicator_impl_neomedia_codec_audio_g722_JNIDecoder_g722_1decoder_1open
     (JNIEnv *jniEnv, jclass clazz)
 {
-    return (jlong) g722_decode_init(NULL, 64000, 0);
+    return (jlong) (intptr_t) g722_decode_init(NULL, 64000, 0);
 }
 
 JNIEXPORT void JNICALL
@@ -38,7 +40,7 @@ Java_net_java_sip_communicator_impl_neomedia_codec_audio_g722_JNIDecoder_g722_1d
         if (inputPtr)
         {
             g722_decode(
-                    (g722_decode_state_t *) decoder,
+                    (g722_decode_state_t *) (intptr_t) decoder,
                     (int16_t *) (outputPtr + outputOffset),
                     (const uint8_t *) (inputPtr + inputOffset),
                     outputLength / (sizeof(int16_t) * 2));
