+++ src/native/g722/net_java_sip_communicator_impl_neomedia_codec_audio_g722_JNIEncoder.c
@@ -8,7 +8,7 @@ JNIEXPORT void JNICALL
 Java_net_java_sip_communicator_impl_neomedia_codec_audio_g722_JNIEncoder_g722_1encoder_1close
     (JNIEnv *jniEnv, jclass clazz, jlong encoder)
 {
-    g722_encode_state_t *e = (g722_encode_state_t *) encoder;
+    g722_encode_state_t *e = (g722_encode_state_t *) (intptr_t) encoder;
 
     g722_encode_release(e);
     g722_encode_free(e);
@@ -18,7 +18,7 @@ JNIEXPORT jlong JNICALL
 Java_net_java_sip_communicator_impl_neomedia_codec_audio_g722_JNIEncoder_g722_1encoder_1open
     (JNIEnv *jniEnv, jclass clazz)
 {
-    return (jlong) g722_encode_init(NULL, 64000, 0);
+    return (jlong) (intptr_t) g722_encode_init(NULL, 64000, 0);
 }
 
 JNIEXPORT void JNICALL
@@ -38,7 +38,7 @@ Java_net_java_sip_communicator_impl_neomedia_codec_audio_g722_JNIEncoder_g722_1e
         if (inputPtr)
         {
             g722_encode(
-                    (g722_encode_state_t *) encoder,
+                    (g722_encode_state_t *) (intptr_t) encoder,
                     (uint8_t *) (outputPtr + outputOffset),
                     (const int16_t *) (inputPtr + inputOffset),
                     2 * (outputLength / sizeof(uint8_t)));
